add set em aluno ,melhorado Interessados

+++ backend/src/main/java/com/example/backend/model/Disciplina.java
@@ -72,7 +72,7 @@ public class Disciplina implements Serializable{
 	}
 
 
-	public float getCodigo_disciplina() {
+	public Integer getCodigo_disciplina() {
 		return codigo_disciplina;
 	}
